Move unmerged files under 'site' directory
