+++ site/scala-exercises-client/src/main/scala/utils/DomHandler.scala
